+++ tests/worker.php
@@ -3,16 +3,11 @@
     require __DIR__ . DIRECTORY_SEPARATOR . 'autoload.php';
     import('net.nosial.tamerlib');
 
-    $bot = new TgBotLib\Bot('bot_token');
+    $bot = new TgBotLib\Bot('<BOT TOKEN>');
 
     $bot->setCommandHandler('start', new \commands\StartCommand());
     $bot->setCommandHandler('hash', new \commands\HashCommand());
 
-    TamerLib\Tamer::initWorker();
-
-    TamerLib\Tamer::addFunction('handle_update', function (\TamerLib\Objects\Job $job) use ($bot)
-    {
-        $bot->handleUpdate(\TgBotLib\Objects\Telegram\Update::fromArray(json_decode($job->getData(), true)));
-    });
-
-    TamerLib\Tamer::work();
+    \TamerLib\tm::initialize(\TamerLib\Enums\TamerMode::WORKER);
+    \TamerLib\tm::addFunction('handle_update', [$bot, 'handleUpdate']);
+    \TamerLib\tm::run();